Fix: separate components and json in to respective folders
